refactor: clean up AI components components from AI page and Editor page
